Adventurer refactoring (#41)
* Add alias for group of 6 party members

* Extract three example adventurers functions

* Change signatures of cave_lore_present and woodsman_present to return bool

* Add unit tests for adventurers

Leverage the Catch2 framework

* Re-org test a little to please Visual Studio et al

Ripple the enable_tests up to the top level so that it can be recognised by a run of CTest at any level (that was complicated to work out!).
Also add to and improve the quality of the adventurer tests.

* Leverage C++ algorithms in adventurer code

* Remove unreferenced function take_sp

* Change calling pattern for heal_party and heal_pc

This will be a general adjustment where we move from relying on a global and index instead preferring a passed in direct value.

* Move heal_party to adventurers.cpp and heal_pc to pc.cpp

* Add test for heal_party, and added comparison operators to all game structs to make testing easier (amongst other things)

* Tidy up heal_party now

The check for status was redundant as there is a check for this in heal_pc

* Add heal_pc test and simplify function

* Adjust cure_pc/cure_party to move away from globals

* Pull out innards of cure_pc into pc_cure and cure_party into advanturers_cure

Also add some tests.

* C++'ize adventurers_cure

Remove main_status check as this is taken care of in pc_cure function

* Rename heal_pc to pc_heal

* Rename heal_party to party_heal

* Change restore_sp_pc and restore_sp_party to pass state rather than use globals

* Move restore_sp_pc and restore_sp_party to new homes

* Rename functions according to their new location

* Add unit tests and simplify functions

* More CMake changes to try and clarify/tidy-up warnings

* Provide option to disable building of tests and use it to stop upsetting Microsoft static analysis

+++ src/utilities/Dump Scenario/dump_scenario.cpp
@@ -59,7 +59,7 @@ bool dump_scenario(std::ostream_iterator<char> strm, std::istream & input)
 		std::format_to(strm, "scen_str[{}]: ({}) \"{}\"\n", index, len, storage.data());
 	}
 
-	for (size_t outdoors_index = 0; outdoors_index < (scenario.out_width * scenario.out_height); ++outdoors_index)
+	for (int outdoors_index = 0; outdoors_index < (scenario.out_width * scenario.out_height); ++outdoors_index)
 	{
 		std::format_to(strm, "Start position: {} {:x}\n", outdoors_index, static_cast<std::size_t>(input.tellg()));
 		outdoor_record_type outdoors;
